+++ xtask/src/commands/all.rs
@@ -2,6 +2,7 @@ use anyhow::Result;
 use structopt::StructOpt;
 
 use super::Compliance;
+use super::Licenses;
 use super::Lint;
 use super::Test;
 
@@ -9,6 +10,8 @@ use super::Test;
 pub struct All {
     #[structopt(flatten)]
     compliance: Compliance,
+    #[structopt(flatten)]
+    licenses: Licenses,
     #[structopt(flatten)]
     lint: Lint,
     #[structopt(flatten)]
@@ -18,7 +21,9 @@ pub struct All {
 impl All {
     pub fn run(&self) -> Result<()> {
         eprintln!("Checking licenses...");
-        self.compliance.run_local()?;
+        self.licenses.run_local()?;
+        eprintln!("Checking compliance...");
+        self.compliance.run()?;
         eprintln!("Checking format and clippy...");
         self.lint.run_local()?;
         eprintln!("Running tests...");
