+++ xtask/src/commands/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod all;
 pub(crate) mod compliance;
 pub(crate) mod dist;
+pub(crate) mod licenses;
 pub(crate) mod lint;
 pub(crate) mod package;
 pub(crate) mod release;
@@ -9,6 +10,7 @@ pub(crate) mod test;
 pub(crate) use all::All;
 pub(crate) use compliance::Compliance;
 pub(crate) use dist::Dist;
+pub(crate) use licenses::Licenses;
 pub(crate) use lint::Lint;
 pub(crate) use package::Package;
 pub(crate) use test::Test;
